a version that technically work
you can't add a bunch of dots to a shape in an svg file. So we
were populating all the circles in a place they couldn't really
go. Instead, this adds the rectangle to a group, and then adds the
circles to the same group, which allows everything to show up